requirements.txt
google-adk
fastapi
uvicorn
python-dotenv
pydantic
requests
langchain_huggingface
slack-bolt
slack_sdk